remove the need for token_index, since it doesn't work with import
I'm surprised of how long it worked
and I was knowingly doing it "wrong" (probably was faster) to see how long it could last
I forgot about it too though.
Without this, imported expressions that need to acces their tokens may
not work

REM we should also avoid the Expr having a token, and store the line and
col in each of them, and whatever else may be necessary.
But no Token's after parse anymore, please

This is only an intermediary refactor towards the future